LFS: Faster autorotate and bytes operations
Unity render is horizontally swapped so we need to mirror swap on X:
that part was rewritten so it's simplified and optimized.

Also if autorotate previous code was doing byte wizzardry but that's not
necessary and slow. In this new version we put the camera so it's always
upright then renders, so it's already good.